refactor(schema): move atropa types out of the server folder
