+++ packages/atropa/src/schema/atropa/block/template.d.ts
@@ -4,18 +4,18 @@ import type {
 	BlockEventsFunction,
 	BlockPermutationsFunction
 } from './functions'
-import type { Namespace } from '../../common/template'
-import type { UseFunction } from '../../common/functions'
-import type { BlockComponents_1_16_0 } from '../../../vanilla/block/v1.16.0'
+import type { Namespace } from '../common/template'
+import type { UseFunction } from '../common/functions'
+import type { BlockComponents_1_16_0 } from '../../vanilla/block/v1.16.0'
 import type {
 	BlockComponents_1_16_100,
 	BlockEventResponses_1_16_100
-} from '../../../vanilla/block/v1.16.100'
-import type { BlockComponents_1_18_10 } from '../../../vanilla/block/v1.18.10'
-import type { BlockComponents_1_18_30 } from '../../../vanilla/block/v1.18.30'
-import type { BlockComponents_1_19_10 } from '../../../vanilla/block/v1.19.10'
-import type { BlockComponents_1_19_20 } from '../../../vanilla/block/v1.19.20'
-import type { Randomize, Sequence } from '../../../vanilla/event/common'
+} from '../../vanilla/block/v1.16.100'
+import type { BlockComponents_1_18_10 } from '../../vanilla/block/v1.18.10'
+import type { BlockComponents_1_18_30 } from '../../vanilla/block/v1.18.30'
+import type { BlockComponents_1_19_10 } from '../../vanilla/block/v1.19.10'
+import type { BlockComponents_1_19_20 } from '../../vanilla/block/v1.19.20'
+import type { Randomize, Sequence } from '../../vanilla/event/common'
 
 interface BlockTemplate_1_16_0
 	extends BlockDescriptionFunction<false>,
