+++ packages/atropa/src/schema/atropa/component/index.d.ts
