+++ packages/atropa/src/schema/atropa/component/template.d.ts
