+++ packages/atropa/src/schema/atropa/item/template.d.ts
