+++ packages/atropa/src/schema/atropa/loot-table/index.d.ts
