+++ packages/atropa/src/schema/atropa/loot-table/template.d.ts
