+++ packages/atropa/src/schema/atropa/recipe/template.d.ts
